Moved the HydroPowerRecorder in to pywr.recorders.
